refactor: pipeline run and release plan list views
