+++ src/components/IntegrationTests/IntegrationTestDetails/tabs/IntegrationTestPipelineRunTab.tsx
@@ -1,6 +1,12 @@
 import * as React from 'react';
 import { useParams } from 'react-router-dom';
 import { Bullseye, Spinner, Title } from '@patternfly/react-core';
+import {
+  INTEGRATION_TEST_PIPELINE_RUN_COLUMNS_DEFINITIONS,
+  DEFAULT_VISIBLE_PIPELINE_RUN_COLUMNS_NO_VULNERABILITIES,
+  NON_HIDABLE_PIPELINE_RUN_COLUMNS,
+  PipelineRunColumnKeys,
+} from '../../../../consts/pipeline';
 import { PipelineRunLabel } from '../../../../consts/pipelinerun';
 import { useLocalStorage } from '../../../../hooks/useLocalStorage';
 import { usePipelineRuns } from '../../../../hooks/usePipelineRuns';
@@ -13,12 +19,6 @@ import { useNamespace } from '../../../../shared/providers/Namespace';
 import { PipelineRunKind } from '../../../../types';
 import { BaseTextFilterToolbar } from '../../../Filter/toolbars/BaseTextFIlterToolbar';
 import PipelineRunEmptyState from '../../../PipelineRun/PipelineRunEmptyState';
-import {
-  INTEGRATION_TEST_PIPELINE_RUN_COLUMNS_DEFINITIONS,
-  DEFAULT_VISIBLE_PIPELINE_RUN_COLUMNS_NO_VULNERABILITIES,
-  NON_HIDABLE_PIPELINE_RUN_COLUMNS,
-  PipelineRunColumnKeys,
-} from '../../../PipelineRun/PipelineRunListView/pipelinerun-columns-config';
 import { getPipelineRunListHeader } from '../../../PipelineRun/PipelineRunListView/PipelineRunListHeader';
 import { PipelineRunListRowWithColumns } from '../../../PipelineRun/PipelineRunListView/PipelineRunListRow';
 import { IntegrationTestLabels } from '../../IntegrationTestForm/types';
